feat: minor code improvements
